Mirror Label in front + fixed most warnings
Mirror label has been put in front of the overlay. This has been done using the Ordering tab that the label has and updating the Z index to be higher.

Made sure that most warnings are gone. There's only 1 warning left, but the solution I found only reproduced the same error. When removing the code, the game worked as intentioned. Is the code needed?